+++ src/Sulu/Bundle/ContactBundle/Controller/TemplateController.php
@@ -114,12 +114,14 @@ class TemplateController extends RestController
             'emailTypes' => $values['emailTypes'],
             'urlTypes' => $values['urlTypes'],
             'faxTypes' => $values['faxTypes'],
+            'socialMediaProfileTypes' => $values['socialMediaProfileTypes'],
             'countries' => $values['countries'],
             'defaultPhoneType' => $defaults['phoneType'],
             'defaultEmailType' => $defaults['emailType'],
             'defaultAddressType' => $defaults['addressType'],
             'defaultUrlType' => $defaults['urlType'],
             'defaultFaxType' => $defaults['faxType'],
+            'defaultSocialMediaProfileType' => $defaults['socialMediaProfileType'],
             'defaultCountry' => $defaults['country'],
         ];
     }
@@ -156,6 +158,10 @@ class TemplateController extends RestController
             ->getRepository('SuluContactBundle:FaxType')
             ->findAll();
 
+        $values['socialMediaProfileTypes'] = $this->getDoctrine()
+            ->getRepository('SuluContactBundle:SocialMediaProfileType')
+            ->findAll();
+
         $values['countries'] = $this->getDoctrine()
             ->getRepository('SuluContactBundle:Country')
             ->findAll();
@@ -178,29 +184,28 @@ class TemplateController extends RestController
             ->getRepository($emailTypeEntity)
             ->find($config['emailType']);
 
-        $phoneTypeEntity = 'SuluContactBundle:PhoneType';
         $defaults['phoneType'] = $this->getDoctrine()
-            ->getRepository($phoneTypeEntity)
+            ->getRepository('SuluContactBundle:PhoneType')
             ->find($config['phoneType']);
 
-        $addressTypeEntity = 'SuluContactBundle:AddressType';
         $defaults['addressType'] = $this->getDoctrine()
-            ->getRepository($addressTypeEntity)
+            ->getRepository('SuluContactBundle:AddressType')
             ->find($config['addressType']);
 
-        $urlTypeEntity = 'SuluContactBundle:UrlType';
         $defaults['urlType'] = $this->getDoctrine()
-            ->getRepository($urlTypeEntity)
+            ->getRepository('SuluContactBundle:UrlType')
             ->find($config['urlType']);
 
-        $faxTypeEntity = 'SuluContactBundle:FaxType';
         $defaults['faxType'] = $this->getDoctrine()
-            ->getRepository($faxTypeEntity)
+            ->getRepository('SuluContactBundle:FaxType')
             ->find($config['faxType']);
 
-        $countryEntity = 'SuluContactBundle:Country';
+        $defaults['socialMediaProfileType'] = $this->getDoctrine()
+            ->getRepository('SuluContactBundle:SocialMediaProfileType')
+            ->find($config['socialMediaProfileType']);
+
         $defaults['country'] = $this->getDoctrine()
-            ->getRepository($countryEntity)
+            ->getRepository('SuluContactBundle:Country')
             ->findOneByCode($config['country']);
 
         return $defaults;